requirements.txt
Pillow
numpy
aiohttp
